test(architecture): move api contract observability tests
